:sparkles: 카테고리 user가 가게 소유주인지 확인하는 로직 추가

+++ src/main/kotlin/com/grapefruit/aid/domain/category/service/impl/DeleteCategoryServiceImpl.kt
@@ -3,6 +3,8 @@ package com.grapefruit.aid.domain.category.service.impl
 import com.grapefruit.aid.domain.category.exception.CategoryNotFoundException
 import com.grapefruit.aid.domain.category.repository.CategoryRepository
 import com.grapefruit.aid.domain.category.service.DeleteCategoryService
+import com.grapefruit.aid.domain.store.exception.UserMismatchException
+import com.grapefruit.aid.domain.user.util.UserUtil
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -10,10 +12,13 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 @Transactional(rollbackFor = [Exception::class])
 class DeleteCategoryServiceImpl(
-    private val categoryRepository: CategoryRepository
+    private val categoryRepository: CategoryRepository,
+    private val userUtil: UserUtil
 ): DeleteCategoryService {
     override fun execute(categoryId: Long) {
         val category = categoryRepository.findByIdOrNull(categoryId) ?: throw CategoryNotFoundException()
+        if(category.store.user != userUtil.currentUser())
+            throw UserMismatchException()
         categoryRepository.delete(category)
     }
